+++ Yotalab.PlanningPoker.Api/Program.cs
@@ -1,32 +1,19 @@
-using Microsoft.AspNetCore.Hosting;
+using System;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Orleans;
 using Orleans.Hosting;
 using Yotalab.PlanningPoker.Hosting;
 
-namespace Yotalab.PlanningPoker.Api
-{
-  public class Program
-  {
-    public static void Main(string[] args)
-    {
-      CreateHostBuilder(args).Build().Run();
-    }
+Console.Title = "PlanningPoker API";
 
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
-      Host.CreateDefaultBuilder(args)
-      .ConfigureWebHostDefaults(webBuilder =>
-      {
-        webBuilder.UseStartup<Startup>();
-      })
-      .ConfigureServices(services =>
-      {
-        services.Configure<ConsoleLifetimeOptions>(options =>
-        {
-          options.SuppressStatusMessages = true;
-        });
-      })
-      .UseOrleansSiloInProcess();
-  }
-}
+await Host.CreateDefaultBuilder()
+  .UseOrleansSiloInProcess()
+  .ConfigureServices(services =>
+  {
+    services.Configure<ConsoleLifetimeOptions>(options =>
+    {
+      options.SuppressStatusMessages = true;
+    });
+  })
+  .RunConsoleAsync();
