+++ Yotalab.PlanningPoker.Api/appsettings.Development.json
@@ -1,9 +1,48 @@
 {
-  "Logging": {
-    "LogLevel": {
+  "DetailedErrors": true,
+  "Serilog": {
+    "MinimumLevel": {
       "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
+      "Override": {
+        "Microsoft": "Warning"
+      }
+    },
+    "Enrich": [ "FromLogContext", "WithThreadId" ],
+    "WriteTo": [
+      {
+        "Name": "Async",
+        "Args": {
+          "configure": [
+            {
+              "Name": "File",
+              "Args": {
+                "path": "%LOGS_DIRECTORY%/orleans.log",
+                "outputTemplate": "{Timestamp:o} [{Level:u3}] ({ThreadId}) ({SourceContext}) {Message}{NewLine}{Exception}",
+                "rollingInterval": "Day",
+                "fileSizeLimitBytes": 7340032,
+                "rollOnFileSizeLimit": true,
+                "retainedFileCountLimit": 14
+              }
+            }
+          ]
+        }
+      }
+    ],
+    "Filter": [
+      {
+        "Name": "ByExcluding",
+        "Args": {
+          "expression": "StartsWith(SourceContext, 'Orleans.')"
+        }
+      }
+    ]
+  },
+  "ConnectionStrings": {
+    "DefaultGrainStorage": "Server=localhost;Port=3306;Database=agilepoker;User=root;Password=11111",
+    "DefaultPubSubStorage": "Server=localhost;Port=3306;Database=agilepoker;User=root;Password=11111"
+  },
+  "Orleans": {
+    "SiloPort": 11111,
+    "GatewayPort": 30000
   }
 }
