+++ Yotalab.PlanningPoker.Api/appsettings.json
@@ -1,61 +1,4 @@
 {
-  "Serilog": {
-    "MinimumLevel": {
-      "Default": "Debug",
-      "Override": {
-        "Microsoft": "Warning"
-      }
-    },
-    "WriteTo:Async": {
-      "Name": "Async",
-      "Args": {
-        "configure": [
-          {
-            "Name": "File",
-            "Args": {
-              "path": "..\\logs\\server.log",
-              "outputTemplate": "{Timestamp:o} [{Level:u3}] ({ThreadId}) {Message}{NewLine}{Exception}",
-              "rollingInterval": "Day",
-              "fileSizeLimitBytes": 7340032,
-              "rollOnFileSizeLimit": true,
-              "retainedFileCountLimit": 14
-            }
-          }
-        ]
-      }
-    },
-    "WriteTo:Debug": {
-      "Name": "Debug",
-      "Args": {
-        "configure": [
-          {
-            "Name": "Debug",
-            "Args": {
-              "outputTemplate": "{Timestamp:o} [{Level:u3}] ({ThreadId}) {Message}{NewLine}{Exception}"
-            }
-          }
-        ]
-      }
-    },
-    "WriteTo:Console": {
-      "Name": "Debug",
-      "Args": {
-        "configure": [
-          {
-            "Name": "Debug",
-            "Args": {
-              "outputTemplate": "{Timestamp:o} [{Level:u3}] ({ThreadId}) {Message}{NewLine}{Exception}"
-            }
-          }
-        ]
-      }
-    }
-  },
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost;Port=3306;Database=agilepoker;User=root;Password=11111",
-    "DefaultGrainStorage": "Server=localhost;Port=3306;Database=agilepoker;User=root;Password=11111",
-    "DefaultPubSubStorage": "Server=localhost;Port=3306;Database=agilepoker;User=root;Password=11111"
-  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
