+++ Yotalab.PlanningPoker.BlazorServerSide/GenericHostExtensions.cs
@@ -3,25 +3,25 @@ using System.Net.NetworkInformation;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Orleans;
-using Orleans.Configuration;
 using Orleans.Hosting;
 using Serilog;
-using Yotalab.PlanningPoker.Grains;
-using Yotalab.PlanningPoker.Grains.Interfaces;
 
 namespace Yotalab.PlanningPoker.Hosting
 {
   public static class GenericHostExtensions
   {
+    public static IHostBuilder UseOrleansClient(this IHostBuilder hostBuilder)
+    {
+      return hostBuilder;
+    }
+
     public static IHostBuilder UseOrleansSiloInProcess(this IHostBuilder hostBuilder)
     {
       return hostBuilder
         .UseSerilog((context, configuration) => configuration.ReadFrom.Configuration(context.Configuration))
         .UseOrleans((context, builder) =>
         {
-          builder.ConfigureApplicationParts(manager => manager
-              .AddApplicationPart(typeof(SessionGrain).Assembly)
-              .AddApplicationPart(typeof(ISessionGrain).Assembly));
+          builder.ConfigureApplicationParts(manager => manager.AddFromApplicationBaseDirectory());
 
           // При локальном рамещении нет возможности настроить Orleans, чтобы он не занимал порты.
           // Issue запланированна на 4.0 версию https://github.com/dotnet/orleans/issues/7023
@@ -31,28 +31,12 @@ namespace Yotalab.PlanningPoker.Hosting
           var siloPort = context.Configuration.GetValue<int>("Orleans:SiloPort", 11111);
           var gatewayPort = context.Configuration.GetValue<int>("Orleans:GatewayPort", 30000);
 
-          var clusterConnectionString = Environment.GetEnvironmentVariable("CLUSTER_STORAGE");
-          if (string.IsNullOrWhiteSpace(clusterConnectionString))
-            builder.UseLocalhostClustering(siloPort, gatewayPort);
-          else
-            builder.UseAdoNetClustering(options =>
-            {
-              options.Invariant = "MySql.Data.MySqlConnector";
-              options.ConnectionString = clusterConnectionString;
-            })
-            .Configure<ClusterOptions>(options =>
-            {
-              options.ClusterId = "planingpoker-cluster";
-              options.ServiceId = "planingpoker";
-            })
-            .ConfigureEndpoints(siloPort, gatewayPort);
+          builder.UseLocalhostClustering(siloPort, gatewayPort);
 
           builder.AddAdoNetGrainStorageAsDefault(options =>
           {
             options.Invariant = "MySql.Data.MySqlConnector";
-            options.ConnectionString =
-              Environment.GetEnvironmentVariable("GRAIN_STORAGE") ??
-              context.Configuration.GetConnectionString("DefaultGrainStorage");
+            options.ConnectionString = context.Configuration.GetConnectionString("DefaultGrainStorage");
             options.UseJsonFormat = true;
             options.ConfigureJsonSerializerSettings = (jsonOptions) =>
             {
@@ -63,16 +47,12 @@ namespace Yotalab.PlanningPoker.Hosting
           builder.AddAdoNetGrainStorage("PubSubStore", options =>
           {
             options.Invariant = "MySql.Data.MySqlConnector";
-            options.ConnectionString =
-              Environment.GetEnvironmentVariable("PUB_SUB_STORAGE") ??
-              context.Configuration.GetConnectionString("DefaultPubSubStorage");
+            options.ConnectionString = context.Configuration.GetConnectionString("DefaultPubSubStorage");
             options.UseJsonFormat = true;
           });
+
+
           builder.AddSimpleMessageStreamProvider("SMS");
-          builder.UseDashboard(options =>
-          {
-            options.HideTrace = true;
-          });
         });
     }
