+++ Yotalab.PlanningPoker.BlazorServerSide/appsettings.Development.json
@@ -23,7 +23,7 @@
                     {
                       "Name": "File",
                       "Args": {
-                        "path": "..\\logs\\mailing.log",
+                        "path": "%LOGS_DIRECTORY%/mailing.log",
                         "outputTemplate": "{Timestamp:o} {Message}{NewLine}{Exception}",
                         "rollingInterval": "Day",
                         "fileSizeLimitBytes": 7340032,
@@ -59,7 +59,7 @@
                     {
                       "Name": "File",
                       "Args": {
-                        "path": "..\\logs\\server.log",
+                        "path": "%LOGS_DIRECTORY%/server.log",
                         "outputTemplate": "{Timestamp:o} [{Level:u3}] ({ThreadId}) {Message}{NewLine}{Exception}",
                         "rollingInterval": "Day",
                         "fileSizeLimitBytes": 7340032,
@@ -96,7 +96,7 @@
     "DefaultPubSubStorage": "Server=localhost;Port=3306;Database=agilepoker;User=root;Password=11111"
   },
   "DataProtection": {
-    "Directory": "..\\keys",
+    "Directory": "../keys",
     "CertificateThumbprint": ""
   },
   "SmtpEmailSender": {
