+++ Yotalab.PlanningPoker.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Yotalab.PlanningPoker.Blazo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Yotalab.PlanningPoker.Grains", "Yotalab.PlanningPoker.Grains\Yotalab.PlanningPoker.Grains.csproj", "{6E8BB09B-EE37-47FC-9326-674C04CD3797}"
 EndProject
+Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{1913029A-039A-474F-8EA5-02409AB396EE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -51,6 +53,14 @@ Global
 		{6E8BB09B-EE37-47FC-9326-674C04CD3797}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6E8BB09B-EE37-47FC-9326-674C04CD3797}.Release|x86.ActiveCfg = Release|x86
 		{6E8BB09B-EE37-47FC-9326-674C04CD3797}.Release|x86.Build.0 = Release|x86
+		{1913029A-039A-474F-8EA5-02409AB396EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1913029A-039A-474F-8EA5-02409AB396EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1913029A-039A-474F-8EA5-02409AB396EE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1913029A-039A-474F-8EA5-02409AB396EE}.Debug|x86.Build.0 = Debug|Any CPU
+		{1913029A-039A-474F-8EA5-02409AB396EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1913029A-039A-474F-8EA5-02409AB396EE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1913029A-039A-474F-8EA5-02409AB396EE}.Release|x86.ActiveCfg = Release|Any CPU
+		{1913029A-039A-474F-8EA5-02409AB396EE}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
